LibWeb: Split TextDecorationStyleValue out of StyleValue.{h,cpp}
